spec_arrow_read_table_arrow

spec_arrow_read_table_arrow

Value

dbReadTableArrow() returns a data frame that contains the complete data from the remote table, effectively the result of calling dbGetQuery()

with SELECT * FROM <name>.

An empty table is returned as a data frame with zero rows.

Failure modes

An error is raised if the table does not exist.

An error is raised when calling this method for a closed or invalid connection. An error is raised if name cannot be processed with dbQuoteIdentifier() or if this results in a non-scalar.

Specification

The name argument is processed as follows, to support databases that allow non-syntactic names for their objects:

If an unquoted table name as string: dbReadTableArrow() will do the quoting, perhaps by calling dbQuoteIdentifier(conn, x = name)

If the result of a call to dbQuoteIdentifier(): no more quoting is done

See also