spec_driver_get_info

spec_driver_get_info

spec_connection_get_info

spec_meta_get_info_result

Value

For objects of class DBIDriver, dbGetInfo() returns a named list that contains at least the following components:

driver.version: the package version of the DBI backend,

client.version: the version of the DBMS client library.

For objects of class DBIConnection, dbGetInfo() returns a named list that contains at least the following components:

db.version: version of the database server,

dbname: database name,

username: username to connect to the database,

host: hostname of the database server,

port: port on the database server. It must not contain a password component. Components that are not applicable should be set to NA.

For objects of class DBIResult, dbGetInfo() returns a named list that contains at least the following components:

statatment: the statement used with dbSendQuery() or dbExecute(), as returned by dbGetStatement(),

row.count: the number of rows fetched so far (for queries), as returned by dbGetRowCount(),

rows.affected: the number of rows affected (for statements), as returned by dbGetRowsAffected()

has.completed: a logical that indicates if the query or statement has completed, as returned by dbHasCompleted().