spec_meta_get_statement

spec_meta_get_statement

Value

dbGetStatement() returns a string, the query used in either dbSendQuery()

or dbSendStatement().

Failure modes

Attempting to query the statement for a result set cleared with dbClearResult() gives an error.

See also