spec_sql_list_tables

spec_sql_list_tables

Value

dbListTables() returns a character vector that enumerates all tables and views in the database. Tables added with DBI::dbWriteTable() are part of the list. As soon a table is removed from the database, it is also removed from the list of database tables.

The same applies to temporary tables if supported by the database.

The returned names are suitable for quoting with dbQuoteIdentifier().

Failure modes

An error is raised when calling this method for a closed or invalid connection.

See also