Test the "Connection" class

Test the "Connection" class

Usage

test_connection(skip = NULL, run_only = NULL, ctx = get_default_context())

Arguments

skip

[character()]
A vector of regular expressions to match against test names; skip test if matching any. To improve precision, the regular expressions are matched against the entire test name minus a possible suffix _N where N is a number. For example, skip = "exists_table" will skip both "exists_table_1" and "exists_table_2", but not "there_exists_table".

run_only

[character()]
A vector of regular expressions to match against test names; run only these tests. The regular expressions are matched against the entire test name.

ctx

[DBItest_context]
A test context as created by make_context().

See also