Test the "Driver" class

Test the "Driver" class

test_driver(skip = NULL, run_only = NULL, ctx = get_default_context())

Arguments

skip
[character()]
A vector of regular expressions to match against test names; skip test if matching any. The regular expressions are matched against the entire test name.

run_only
[character()]
A vector of regular expressions to match against test names; run only these tests. The regular expressions are matched against the entire test name.

ctx
[DBItest_context]
A test context as created by make_context().

See also